03062014 SME with new QR Code reader with no live testing

+++ ttttt/HeaderView.m
@@ -34,13 +34,13 @@
      NSDictionary *dict = [NSDictionary dictionaryWithObject:Headerviewlabel.text forKey:@"Searchterm"];
     switch (index) {
         case 0:
-            NSLog(@"%s item:%d", __PRETTY_FUNCTION__, index);
+            NSLog(@"%s item:%ld", __PRETTY_FUNCTION__, (long)index);
             [[NSNotificationCenter defaultCenter] postNotificationName:@"TestNotification" object:self userInfo:nil];
           
             break;
         case 1:
            
-            NSLog(@"%s item:%d", __PRETTY_FUNCTION__, index);
+            NSLog(@"%s item:%ld", __PRETTY_FUNCTION__, (long)index);
             [[NSNotificationCenter defaultCenter]postNotificationName:@"Searchactivated" object:self userInfo:dict];
             break;
